Everything should now be moved to the fuzz_main.rs file
